pyproject.toml
[tool.poetry]
name = "pet-tracker-backend"
version = "0.1.0"
description = "Google Calendar-like backend with FastAPI and PostgreSQL"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "app", from = "." }
]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
python-jose = "^3.4.0"
passlib = "^1.7.4"
python-multipart = "^0.0.20"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.39"
pydantic = {extras = ["email"], version = "^2.10.6"}
sqlmodel = "^0.0.24"
bcrypt = "^4.3.0"
